Fixed the bugs in MagVectorCube but kept the same 3 axes configuration for the data as before. In the end I could not make sense of adding an axis for each magnetic field component.
Added a function make_def_wcs that creates a default wcs object.

Fixed the HMI_data_load notebook to include a function that finds
the nearest HMI data series to an input date and downloads it or
searches for it if the files already exist. This creates a folder
either at user specification or by default to store the data. However,
I modified the .gitignore file to include any default folder that
is named Data/.