Modified a number of methods in StokesCube to improve performance and pass WCS axes consistently between objects.
The _stokes_map() method now generates either a StokesParamCube object or
a StokeParamMap object depending on whether a range of wavelenghts is
provided by setting the stop_wavelenght keyword, or a single wavelength is
entered. If a range of wavelengths is wanted the WCS is also modified to
the new wavelength range.

Implemented an early version of the slide bar to scroll through the
wavelengths in StokesParamCube.